Modifed GameState.cs, Ship.cs, ChooseShip.cs to work better together at setting the shiptype.
LaserGun.cs now checks if the player has control before drawing crosshair and firing laser.

Made a Ship only superiority prefab so that opening menu doesn't use the full ship prefab.

Renamed the other ship only prefabs.

Made LevelStart.cs that loads the appropriate ship prefab into the scene for the use to control.

Made Launch.cs to load the level from the opening menu.

BUGS:

Ship is not taking damage on collisions with the rings.

Opening menu, when a ship button is clicked twice in a row the ship does not spin. This could be a "feature".